refactor: move theme switch button inside core folder
